+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix MenuItem::add() and insert().
 - Add experimental GlWidgetWindow, gated under the enable-glwindow feature.
 - Add app::set_color_with_alpha(), gated under the enable-glwindow feature.
+- Update ttf-parser to 0.15.
 
 ## [1.2.31] - 2022-02-19
 - Add dialog::message_icon().